requirements.txt
alabaster==0.7.12
appnope==0.1.0
asn1crypto==1.3.0
Babel==2.8.0
backcall==0.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
cycler==0.10.0
decorator==4.4.1
docutils==0.14
graphviz==0.13.2
humanize==1.0.0
idna==2.9
imagesize==1.2.0
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.3
neat-python>=0.92
numpy==1.18.1
packaging==20.1
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.3
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.5.2
pyOpenSSL==19.1.0
pyparsing==2.4.6
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
snowballstemmer==1.2.1
Sphinx==2.4.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.2.0
Theano==1.0.4
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8

pandas~=1.0.3
networkx~=2.4